+++ src/main/java/ru/akhmetov/AutoRepair/repositories/OrdersRepository.java
@@ -8,5 +8,5 @@ import ru.akhmetov.AutoRepair.models.Order;
  * @author Oleg Akhmetov on 26.12.2022
  */
 @Repository
-public interface OrdersRepository extends JpaRepository<Order, Long> {
+public interface OrdersRepository extends JpaRepository<Order, Integer> {
 }
